backend/search-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg[binary]==3.2.9
python-decouple==3.8
django-extensions==3.2.3
requests==2.31.0
celery==5.3.4
redis==5.0.1
Pillow==11.3.0 